requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.28.63
botocore==1.31.63
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.0
cryptography==41.0.2
Django==4.2.3
django-admin-argon-dashboard==1.0.16
django-admin-material-dashboard @ git+https://github.com/app-generator/django-admin-material-dashboard.git@2bbf01c4daa19e310b2ebe384ff6485dd72913f2
django-cors-headers==4.3.0
django-filter==23.3
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-mptt==0.15.0
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.3.0
Faker==19.3.0
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.*
jmespath==1.0.1
liqpay==1.0.1
liqpay-python==1.0
packaging==23.2
Pillow==10.0.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.6
whitenoise==6.5.0

djangorestframework-jwt~=1.11.0
googletrans~=4.0.0rc1
python-dotenv~=1.0.0